requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiomysql==0.2.0
aioredis==2.0.1
aiosqlite==0.17.0
aliyun-python-sdk-core==2.13.36
annotated-types==0.5.0
async-timeout==4.0.3
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dictdiffer==0.9.0
html5tagger==1.3.0
httptools==0.6.0
iso8601==1.1.0
jmespath==0.10.0
multidict==6.0.4
passlib==1.7.4
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyMySQL==1.1.0
pypika-tortoise==0.1.6
pytz==2023.3
PyYAML==6.0.1
sanic==23.6.0
sanic-ext==23.6.0
sanic-routing==23.6.0
tomlkit==0.12.1
tortoise-orm==0.20.0
tracerite==1.1.0
typing_extensions==4.7.1
ujson==5.8.0
uvloop==0.17.0
websockets==11.0.3
